broadcast-management-system/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
firebase-admin==6.2.0
requests==2.31.0
